% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear variables}
\usage{
clear()
clear(remove = "var")
clear(remove = "var*")
clear(remove = c("var1","var2"))
clear(keep = "var")
clear(keep = "var*")
}
\description{
This function is a wrapper for the \code{rm()} function. 
This function takes a pattern, which can include a wildcard (*, asterisk) 
in either the argument \code{remove} or \code{keep} to either remove or keep 
all objects matching a string.
It's possible to input multiple variable names in the \code{remove} argument, 
but it's not (yet) possible for the `keep` argument.
}
