% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_alpha_hex.R
\name{full_alpha_hex}
\alias{full_alpha_hex}
\title{Convert to full alpha hex}
\usage{
full_alpha_hex(color, alpha)
}
\description{
Convert color with transparency to the equivalent color without transparency.
}
\author{
Anthony Reinhard (Twitter: reinhurdler)
}
