% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWER.R
\name{FWER}
\alias{FWER}
\title{Calculate FWER}
\usage{
PrFWER <- FWER(26)
PrFWER <- FWER(n = 72, alpha = 0.01)
}
\arguments{
\item{n}{Number of calculations}

\item{alpha}{Alpha level (default at 0.05)}
}
\description{
Calculate Family-Wise Error Rate (FWER) based on the number of comparisons and the alpha level.
}
