% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_double_violin.R
\name{geom_double_violin}
\alias{geom_double_violin}
\title{Double violin plot}
\usage{
geom_double_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Violin plot split across a binary group. Particularly usefulf for checking distributions across sex,
or comparing control to patient group.
}
