% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{get.full.matrix}
\alias{get.full.matrix}
\title{get.full.matrix

Credit: Tobias Kaufmann}
\usage{
get.full.matrix(vector, nnodes)
}
\description{
get.full.matrix

Credit: Tobias Kaufmann
}
\author{
Tobias Kaufmann
}
