% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{get.lower.half.matrix}
\alias{get.lower.half.matrix}
\title{get.lower.half.matrix

Credit: Tobias Kaufmann}
\usage{
get.lower.half.matrix(vector, nnodes)
}
\description{
get.lower.half.matrix

Credit: Tobias Kaufmann
}
\author{
Tobias Kaufmann
}
