% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{get.upper.half.matrix}
\alias{get.upper.half.matrix}
\title{get.upper.half.matrix}
\usage{
get.upper.half.matrix(vector, nnodes)
}
\description{
Matrix modifications
}
\details{
Performs matrix modifications:
get.full.matrix()
get.upper.half.matrix()
get.lower.half.matrix

Credit: Tobias Kaufmann
}
\author{
Daniel Roelfs
}
