% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{get_full_matrix}
\alias{get_full_matrix}
\title{Get full matrix

Credit: Tobias Kaufmann}
\usage{
get_full_matrix(vector, nnodes)
}
\description{
Get full matrix

Credit: Tobias Kaufmann
}
\author{
Tobias Kaufmann
}
