% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{get_lower_half_matrix}
\alias{get_lower_half_matrix}
\title{Get lower half of matrix

Credit: Tobias Kaufmann}
\usage{
get_lower_half_matrix(vector, nnodes)
}
\description{
Get lower half of matrix

Credit: Tobias Kaufmann
}
\author{
Tobias Kaufmann
}
