% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers.R
\name{get_outliers}
\alias{get_outliers}
\title{Get outliers}
\usage{
get_outliers(vector = NULL)
}
\description{
Get the outliers from a vector. Uses the \code{boxplot()} function to detect
outliers. To remove outliers from a vector, use indexing using the output
from this function or use the \code{remove_outliers()} function.
}
