% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{get_upper_half_matrix}
\alias{get_upper_half_matrix}
\title{Get upper half of matrix}
\usage{
get_upper_half_matrix(vector, nnodes)
}
\description{
Matrix modifications
}
\details{
Performs matrix modifications:\cr
\code{get.full.matrix()}\cr
\code{get.upper.half.matrix()}\cr
\code{get.lower.half.matrix}\cr

Credit: Tobias Kaufmann
}
\author{
Daniel Roelfs
}
