% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norment_colors.R
\docType{data}
\name{norment_colors}
\alias{norment_colors}
\title{Get NORMENT colors}
\format{An object of class \code{character} of length 11.}
\usage{
norment_colors
norment_colours
}
\description{
This script contains definitions of the NORMENT colors. Colors are taken
directly from the NORMENT house design document. It also includes colors used
in the NORMENT logo and some other useful colors for greyscale documents that
fit well with the NORMENT theme also. The colors can be indexed in the list,
but can also be used as a non-named string by using:
\code{norment.<color>}
}
\keyword{datasets}
