% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{notin}
\alias{notin}
\alias{\%!in\%}
\title{Negation of the infix 'in' operator}
\usage{
x \%!in\% y
}
\description{
This function is the opposite of the "%in%" operator. It works the same, and
can be used in the same situations. It selects all cases not in another
vector. Output is TRUE or FALSE, never NA.
}
