% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTOPdemographics.R
\name{readTOPdemographics}
\alias{readTOPdemographics}
\title{Import TOP demographic data}
\usage{
TOPdata <- readTOPdemographics()
TOPdata <- readTOPdemographics(file = "<path to file>", demogr_only = TRUE)
}
\arguments{
\item{demogr_only}{TRUE/FALSE whether only key variables should be imported}

\item{file}{file path if different from default cleanfiles}
}
\description{
On TSD, imports the TOP demographics as a data frame, i.e. the CLEANFILES. It 
automatically assigns the correct attributes to some key variables, other 
variables need to be set manually afterwards, e.g. via the `mutate()` 
function.
This function defaults to the setup of TSD, but can be used locally too. In 
that case it will give warnings if the column names are different from what 
it expects. It will detect whether you're working on a Linux, Windows, or
Mac machine and adapt accordingly.
This function as default will import all variables (of which 
there are more than 1000), if you're interested in the key demographic 
variables only, such as age, diagnosis, date of interview, etc., then set the 
`demogr_only` option to `TRUE`. This will then import only about 20 key 
demographic variables.
}
\details{
Disclaimer: this function works only on TSD
}
