% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers from vector}
\usage{
data$x <- remove_outliers(data$x)
data <- data \%>\% mutate(x_no_outliers = remove_outliers(x))
}
\description{
As default, this function replaces the outliers with \code{NA}. Setting the 
\code{replace_with_na} setting to \code{FALSE} will remove the outliers from a vector,
and thus return a shorter vector. Works also within a pipe structure (\code{\%>\%}).
}
