% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_norment.R
\name{scale_fill_norment}
\alias{scale_fill_norment}
\title{ggplot scale_fill element}
\usage{
scale_fill_norment(palette = "default", discrete = TRUE, reverse = FALSE, ...)
}
\description{
Add the NORMENT colors to a ggplot object. This function can be used as one
would always use the "scale_fill_*" objects, except it requires the added
input whether the colors required need to be discrete or continuous, this of
course depends on the data that you're trying to plot. It can also take input
of which of the NORMENT color palettes you want to use. For a list of the
available color palettes type "?norment_palettes". Lastly, you can also tell
this function to reverse the colors in the opposite direction as standard.
}
