% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Calculate Standard Error of the Mean}
\usage{
rSEM <- sem(sample(1:100,50))
rSEM <- sem(c(NA,sample(1:100,50)), na.rm = TRUE)
}
\arguments{
\item{x}{Vector with numeric or integer data}

\item{na.rm}{Remove NA's (TRUE/FALSE)}
}
\description{
Calculate the standard error of the mean (SEM) on a vector of data. Takes
input either with or without NA's, and will then take input as to what to do
with the NA's in the `na.rm` option.
}
