% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_norment_logo.R
\name{show_norment_logo}
\alias{show_norment_logo}
\title{Show the NORMENT logo on your ggplot object}
\usage{
show_norment_logo(
  plot.obj = plot,
  with.name = FALSE,
  pos = c(0.96, 0.05),
  just = "center",
  size = unit(25, "points"),
  ...
)
}
\description{
This function adds the NORMENT logo as a layer on top of your ggplot object.
You can specify whether you want the logo with the name or without. 
Please pay attention that if you use the logo with the name, that you likely have to
reposition the logo with the \code{pos} option. Minimal usage is as follows:
\code{show_norment_logo(ggobj)}.\cr
More options can be set using the \code{with.name} option (\code{TRUE}/\code{FALSE}), 
the \code{pos} option which takes a relative vector for the position. \code{just} to tell
it which justification to assume, and \code{size}, which takes a \code{unit} element to
scale the size of the logo to the desired size.\cr
\cr
This function is not very elegant at the moment, but it might be improved
over time.
}
