% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_norment_palette.R
\name{show_norment_palette}
\alias{show_norment_palette}
\title{Show NORMENT color palettes}
\usage{
show_norment_palette()
show_norment_palette("gradient")
show_norment_palette(c("powerpoint","profile","grey"))
}
\description{
This function shows the NORMENT color palettes as defined in
\code{norment_palette}. It is possible to either plot all by calling the function
without any input arguments, or it's possible to show only the one palette by
inserting the name of the color palette as argument.\cr
NOTE: for the \code{"random"} option, the order of the colors is not shuffled 
in this case. That only happens after calling the \code{norment_pal()} function.
}
