% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_rank.R
\name{signed_rank}
\alias{signed_rank}
\title{Obtain signed rank}
\usage{
x_signed_rank <- signed_rank(x)
}
\arguments{
\item{x}{Vector with numeric or integer data}
}
\description{
Transform a vector of values to its signed rank. This is sometimes necessary
in cases where your data is normally distributed. For instance the Wilcoxon
signed-rank test uses the signed rank of the vector.
}
