% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_corr_matrix.R
\name{simulate_corr_matrix}
\alias{simulate_corr_matrix}
\title{Simulate correlation matrix}
\usage{
corr.matrix <- simulate.corr.matrix(60)
corr.matrix <- simulate.corr.matrix(nnodes = 40, mean = 0, sd = 0.5,
   shape = "upper")
corr.matrix <- simulate.corr.matrix(nnodes = 180, mean = 100, sd = 6,
   shape = "lower")
}
\arguments{
\item{nnodes}{Number of nodes}

\item{mean}{Number of individuals in dataset}

\item{sd}{Add significant SNPs (logical)}

\item{shape}{Shape of the correlation matrix ("full","upper","lower")}
}
\description{
Simulate correlation matrix or any other matrix. Values are normally
distributed around the mean and standard deviation.
}
