% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateGWAS.R
\name{SimulateGWAS}
\alias{SimulateGWAS}
\title{Simulate GWAS data}
\usage{
sim.dat <- SimulateGWAS()
sim.dat <- SimulateGWAS(nSNPs = 5e5, N = 2e4, 
   AddSigSNPs = FALSE, SigCHR = NULL, nSigCols = NULL)
sim.dat <- SimulateGWAS(nSNPs = 2e5, N = 6e5, 
   AddSigSNPs = TRUE, SigCHR = c(2,4,6,7), nSigCols = 4)
}
\arguments{
\item{nSNPs}{N
AddSigSNPs
SigCHR
nSigCols}
}
\description{
This function serves to simulate GWAS data. The output resembles that of the 
PLINK output when generated using the `--assoc` flag for association testing.
The purpose of this function is mainly for debugging and testing of other 
scripts. Default setting is to create one significant column. This can be set 
in the `AddSigSNPs` field (logical). The parameters of the significant column 
are `SigCHR` to specify which chromosome the significant column should appear 
in (integer between 1 and 22), and the number of significant columns  in 
`nSigCols` (integer).
Other parameters are the number of SNPs (`nSNPs`) in the fake dataset, and 
the number of individuals in the dataset. Note that these last two parameters 
are subject to change due to the algorithm by which random data is generated 
in this function.
}
