% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sine.R
\name{sine}
\alias{sine}
\title{Create sine wave}
\usage{
v <- sine(from = -2, to = 2*pi, length = 1000)
v <- sine(-pi, 20, 1e5)
}
\arguments{
\item{from}{single value indicated from where the vector should start}

\item{to}{single value indicated from where the vector should end}

\item{length}{length of the output vector}
}
\description{
Provides a vector with the ouput from a `sin()` function. This function is 
nothng more than a wrapper for said function. It takes a range of the sine 
wave from a specific data point to a specific data point (e.g. -pi to pi).
Another input is the length of the output vector. Useful for cases where you
want to perform a sine transformation on your data.
}
