% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_norment.R
\name{theme_norment}
\alias{theme_norment}
\title{NORMENT theme for ggplot objects}
\usage{
theme_norment(
  base_family = "sans",
  base_size = 9,
  plot_title_family = base_family,
  plot_title_size = 16,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = base_family,
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 9,
  caption_face = "italic",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = 10.5,
  axis_title_face = "plain",
  axis_title_just = "c",
  plot_margin = margin(30, 30, 30, 30),
  grid = TRUE,
  grid_col = NULL,
  axis = FALSE,
  axis_col = NULL,
  ticks = FALSE,
  legend = TRUE,
  bg_col = NULL,
  transparent = FALSE,
  dark = FALSE
)
}
\description{
This is a theme developed for use in combination with the \code{ggplot2} package.
It adds the default NORMENT font (Arial, support for other fonts used by
NORMENT might be added later). There's quite some space for text and
explanation, e.g. title, subtitle, caption, and so on. This theme is highly
adaptable, so many elements in each plot can be formatted to personal taste.
While this function contains the general style of the plot, it does not
automatically use the NORMENT colors on data presented in the plot. In order
to use the NORMENT colors also, please refer to the \code{scale_color_norment} and
the \code{scale_fill_norment} functions, also included in the \code{normentR} package.
}
