% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.list.R
\name{variable.list}
\alias{variable.list}
\title{Get variable list}
\usage{
variable.list()
}
\description{
This function is a wrapper for both the \code{ls()} and the \code{typeof} function.
Output is a data frame with the index, variable name, and the type of all variables 
in the global environment (\code{.GlobalEnv}).
}
