% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear variables}
\usage{
clear()
clear(remove = "var")
clear(remove = "var*")
clear(keep = "var")
clear(keep = "var*")
}
\description{
This function is a wrapper for the `rm()` function. This function takes a 
pattern, which can include a wildcard (*, asterisk) in either the argument 
`remove` or `keep` to either remove or keep all objects matching a string.
It's possible to input multiple variable names in the `remove` argument, but 
it's not (yet) possible for the `keep` argument.
}
