% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norment_palette.R
\docType{data}
\name{norment_palette}
\alias{norment_palette}
\title{Definition of the NORMENT color palettes}
\format{An object of class \code{list} of length 11.}
\usage{
norment_palette
}
\description{
This list includes the default palettes included in the package. There are a
few flavors to choose from, such as "default", which includes the default
colors included in the NORMENT design guide, as well as "logo", which
includes colors featured on the NORMENT logo, and a few more. "purple" and
"blue" include only the purple and blue color variants of the NORMENT colors.
"mixed" uses different colors to maximize color variance in the palette.
"grey" uses only greyscale colors. For a full list of the available colors
type: `available_palettes()`, and to show the palettes type:
`show_norment_palette()`
}
\keyword{datasets}
