% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norment_palette.R
\name{norment_palette}
\alias{norment_palette}
\title{Definition of the NORMENT color palettes}
\description{
This list includes the default palettes included in the package. There are a
few flavors to choose from, such as "default", which includes the default
colors included in the NORMENT design guide, as well as "logo", which
includes colors featured on the NORMENT logo, and a few more. "purple" and
"blue" include only the purple and blue color variants of the NORMENT colors.
"mixed" uses different colors to maximize color variance in the palette.
"grey" uses only greyscale colors. For a full list of the available colors
type: `available_palettes()`, and to show the palettes type:
`show_norment_palette()`
}
\references{
The "scico" palettes included here were developed by Fabio 
Crameri at the University of Oslo, and ported to R by Thomas Pedersen. The 
"ggsci" were developed by Nan Xiao and were developed for his "ggisci" 
package.
}
