% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers from vector}
\usage{
remove_outliers(vector = NULL, replace_with_na = TRUE)
}
\description{
As default, this function replaces the outliers with `NA`. Setting the 
`replace_with_na` setting to `FALSE` will remove the outliers from a vector,
and thus return a shorter vector.
}
