% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_axes.R
\name{shift_axes}
\alias{shift_axes}
\title{Shift axes}
\usage{
shift_axes(plot.obj, x = 0, y = 0)
}
\description{
This functions shifts the axes from the default position to a prespecified
position on the x- or y-axis. This can be particularly useful for plots that
have the null coordinate at x=0 and y=0. This function does not apply itself
to the ggplot object directly, but rather takes the ggplot object as input.
}
\details{
Usage as as follows
shift_axes(ggobj, x = 0, y = 0)
}
