% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_norment_palette.R
\name{show_norment_palette}
\alias{show_norment_palette}
\title{Show NORMENT color palettes}
\usage{
norment_palette()
norment_palette("gradient")
}
\description{
This function shows the NORMENT color palettes as defined in
`norment_palette`. It is possible to either plot all by calling the function
without any input arguments, or it's possible to show only the one palette by
inserting the name of the color palette as argument.
}
